+++ spec/rley/ptree/parse_tree_spec.rb
@@ -2,6 +2,7 @@ require_relative '../../spec_helper'
 
 require_relative '../support/grammar_abc_helper'
 
+
 # Load the class under test
 require_relative '../../../lib/rley/ptree/parse_tree'
 
@@ -9,7 +10,8 @@ module Rley # Open this namespace to avoid module qualifier prefixes
   module PTree # Open this namespace to avoid module qualifier prefixes
     describe ParseTree do
       include GrammarABCHelper  # Mix-in module with builder for grammar abc
-      
+
+
       let(:sample_grammar) do
         builder = grammar_abc_builder
         builder.grammar